+++ symboltable.h
@@ -3,15 +3,25 @@
  * - function declarations regarding the symbol table
  */
 
-#ifndef DEI_STRUCT
-#include "struct.h"
-#endif
+#ifndef SYMBOLTABLE_H_INCLUDED
+#define SYMBOLTABLE_H_INCLUDED
 
-/* -------------- SYMBOL TABLE -------------- */
+#include <string.h> /* needed for strdup() */
+
+#include "struct.h"
+
+/* === DATA === */
 
 #define NHASH 9997 /* size of symbol table */
 struct symbol symtab[NHASH];    /* symbol table itself */
+
+/* === FUNCTIONS === */
+
 struct symbol *lookup(char *s); /* looks up a string in the symbol table and returns the entry */
 
-void printsymtab(void); /* print entries in the table */
+
+/* === MEMORY MANAGEMENT === */
+
 void freesymboltable(void); /* free the entire symbol table */
+
+#endif /* SYMBOLTABLE_H_INCLUDED */
